Update PL models

+++ ForumTask.PL/Models/MessageViewModel.cs
@@ -20,7 +20,7 @@ namespace ForumTask.PL.Models
         public long Id { get; set; }
         public string Text { get; set; }
         public long TopicId { get; set; }
-        public int? AuthorId { get; set; }
+        public long? AuthorId { get; set; }
         public DateTime WriteTime { get; set; }
         public long PositiveCount { get; set; }
         public long NegativeCount { get; set; }
