+++ ForumTask.PL/Models/TopicViewModel.cs
@@ -18,7 +18,7 @@ namespace ForumTask.PL.Models
         public long Id { get; set; }
         public string Title { get; set; }
         public DateTime CreateTime { get; set; }
-        public int? CreatorId { get; set; }
-        public int MessageCount { get; set; }
+        public long? CreatorId { get; set; }
+        public long MessageCount { get; set; }
     }
 }
